raftstore, storage: move storage tests from raftstore to other mods
